Make DictionaryCredentialsProvider fileprivate to Login.swift
